Connected new contract method to Project detail
