+++ pages/project/[objectId].js
@@ -5,11 +5,6 @@ import { useMoralisQuery } from 'react-moralis'
 
 import ProjectDetail from "../../sections/ProjectDetail"
 
-// Blockchain related 
-import donation from '../../abi/donation.json'
-import { useContractRead } from 'wagmi'
-
-
 const Container = styled.div`
   margin-top: 5%;
   margin-bottom: 5%;
@@ -18,9 +13,6 @@ const Container = styled.div`
 const Project = () => {
   const router = useRouter()
   const { objectId } = router.query 
-
-  var microActive = "N/A"
-
   const { data } = useMoralisQuery("Project", (query) => query.equalTo("objectId", objectId));
   const fetchDetail = JSON.parse(
     JSON.stringify(data, [
@@ -39,14 +31,6 @@ const Project = () => {
   const [subcategory, setSubcategory] = useState(null)
   const [pid, setPid] = useState(null)
 
-  const [amPledged, setAmPledged] = useState("N/A")
-  const [amBackers, setAmBackers] = useState("N/A")
-  const [amMicro, setAmMicro] = useState("N/A")
-  const [amDays, setAmDays] = useState("N/A")
-  const [amGoal, setAmGoal] = useState("N/A")
-
-
-
   const getData = async () => {
     try {
       await setTitle(fetchDetail[0].title)
@@ -54,19 +38,6 @@ const Project = () => {
       await setCategory(fetchDetail[0].category)
       await setSubcategory(fetchDetail[0].subcategory)
       await setPid(fetchDetail[0].pid)
-
-      const micros = useContractRead({
-        addressOrName: process.env.NEXT_PUBLIC_AD_DONATOR,
-        contractInterface: donation.abi,
-        functionName: 'getConnectedMicroFunds',
-        chainId: 80001,
-        args: [pid],
-        watch: false,
-      })
-    
-      if (micros.data){
-        microActive = micros.data.toString()
-      }
 
     } catch (error) {
       console.log(error)
@@ -86,13 +57,8 @@ const Project = () => {
           title={title} 
           category={category} 
           subcategory={subcategory} 
-          amBackers={amBackers} 
-          amMicro={amMicro} 
-          amPledged={amPledged} 
-          amDays={amDays} 
-          amGoal={amGoal} 
           image={image} 
-          microActive={microActive} />
+          pid={pid} />
       </Container>
     </>
   )
