+++ sections/start_project/TellStory/index.js
@@ -25,19 +25,16 @@ const TellStory = ({ setStep }) => {
   }
 
   // TBD all fields mandatory
-  // Prefill value with context data 
   // TBD validate HTTPS links
-  // TBD validate - Cannot be ""
-
 
   return (
     <MainContainer>
-      <SectionTitle title="Define your story" subtitle={'TBD some description'} />
+      <SectionTitle title="Define your story" subtitle={'Describe your project and references'} />
       <TellContainer>
         <FormStyle method="post" action="" enctype="multipart/form-data">
           <label className="custom-file-upload">
             <input className="input-file" type="file" name="files[]" id="file" />
-            <span className="input-span">Image Upload (descoped in MVP)</span>
+            <span className="input-span">Image Upload (IPFS TBD)</span>
           </label>
           {story_form.map((sf) => {
             const { title, description, text_field, change, p } = sf;
